Fixed viewport syntax error: changed meta name | content ";" to ","
Fixed error: 
Viewport argument value "device-width;" for key "width" is invalid, and has been ignored. Note that ';' is not a separator in viewport values. The list should be comma-separated.
Viewport argument value "1.0;" for key "initial-scale" was truncated to its numeric prefix. Note that ';' is not a separator in viewport values. The list should be comma-separated.
Viewport argument value "1.0;" for key "maximum-scale" was truncated to its numeric prefix. Note that ';' is not a separator in viewport values. The list should be comma-separated.
"Viewport argument value "no;" for key "user-scalable" is invalid, and has been ignored. Note that ';' is not a separator in viewport values. The list should be comma-separated. "